fixing storybook - for some reason it needs some of `client-apps` dependencies at the top level. will clean up later
