+++ app/client-app/.storybook/preview.js
@@ -1,7 +1,8 @@
 import React from "react";
 import { addDecorator } from "@storybook/react";
 import { withKnobs } from "@storybook/addon-knobs";
-import { ChakraProvider, CSSReset, theme } from "@chakra-ui/core";
+import { ChakraProvider, CSSReset } from "@chakra-ui/core";
+import theme from "@chakra-ui/theme";
 
 addDecorator((s) => (
   <>
