+++ app/client-app/package.json
@@ -2,6 +2,22 @@
   "name": "@decsys/client-app",
   "version": "2.0.0-alpha.1",
   "private": true,
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject",
+    "storybook": "start-storybook -p 9009 -s public"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ],
   "dependencies": {
     "@decsys/param-types": "file:../../packages/param-types",
     "@reach/router": "^1.3.3",
@@ -15,12 +31,10 @@
     "navi": "^0.14.0",
     "react-beautiful-dnd": "^13.0.0",
     "react-busy-indicator": "^1.0.0",
-    "react-dom": "^16.13.1",
     "react-icons": "^3.10.0",
     "react-markdown": "^4.3.1",
     "react-navi": "^0.14.3",
     "react-onclickoutside": "^6.9.0",
-    "react-scripts": "^3.4.1",
     "react-table": "^6.11.5",
     "react-tooltip": "^4.2.6",
     "styled-components": "^5.1.1",
@@ -31,23 +45,7 @@
     "victory": "^34.3.8"
   },
   "devDependencies": {
-    "@storybook/preset-create-react-app": "^1.5.2"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "storybook": "start-storybook -p 9009 -s public",
-    "build-storybook": "build-storybook -s public"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
+    "@storybook/react": "^5.3.19",
+    "react-scripts": "^3.4.1"
+  }
 }
