--- app/client-app/src/stories/components/ComponentEditor/ImageUpload.stories.js
@@ -1,9 +1,0 @@
-import React from "react";
-import ImageUpload from "components/ComponentEditor/ImageUpload";
-
-export default {
-  title: "Admin/ComponentEditor/ImageUpload",
-  component: ImageUpload
-};
-
-export const Basic = () => <ImageUpload params={{}} />;
