+++ app/client-app/stories/components/ComponentEditor/ComponentEditor.stories.js
@@ -1,0 +1,15 @@
+import React from "react";
+import ComponentEditor from "components/ComponentEditor";
+import { paramsLookup, knob, components } from "../ComponentRender.stories";
+
+export default {
+  title: "Admin/ComponentEditor",
+  component: ComponentEditor,
+};
+
+export const Basic = () => (
+  <ComponentEditor
+    component={components[knob()]}
+    params={paramsLookup[knob()].params}
+  />
+);
