+++ package.json
@@ -11,7 +11,8 @@
     "confirmation:build": "rollup -c ./response-items/confirmation/rollup.config.js",
     "confirmation:storybook": "start-storybook -c ./response-items/confirmation/.storybook -p 6006",
     "freetext:storybook": "start-storybook -c ./response-items/freetext/.storybook -p 6006",
-    "freetext:build": "rollup -c ./response-items/freetext/rollup.config.js"
+    "freetext:build": "rollup -c ./response-items/freetext/rollup.config.js",
+    "rating-scales:storybook": "start-storybook -c ./response-items/freetext/.storybook -p 6006"
   },
   "husky": {
     "hooks": {
@@ -35,6 +36,7 @@
     "@storybook/addon-links": "^6.0.0-rc.14",
     "@storybook/addon-storysource": "^6.0.0-rc.14",
     "@storybook/addons": "^6.0.0-rc.14",
+    "@storybook/preset-create-react-app": "^3.1.4",
     "@storybook/react": "^6.0.0-rc.14",
     "@storybook/source-loader": "^6.0.0-rc.14",
     "babel-loader": "^8.1.0",
@@ -51,8 +53,10 @@
     "rollup-plugin-terser": "^6.1.0"
   },
   "dependencies": {
+    "@decsys/client-app": "file:app\\client-app",
     "@decsys/confirmation-response": "file:response-items\\confirmation",
     "@decsys/freetext-response": "file:response-items\\freetext",
-    "@decsys/param-types": "file:packages\\param-types"
+    "@decsys/param-types": "file:packages\\param-types",
+    "@decsys/rating-scales": "file:packages\\rating-scales"
   }
 }
